pyproject.toml
[project]
name = "anima-operating-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "baml-py>=0.89.0",
    "fastapi[standard]>=0.115.12",
    "l>=0.11.0",
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.7",
    "mcp>=1.9.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.41",
    "supabase>=2.15.1",
    "tiktoken>=0.8.0",
]

[tool.setuptools]
packages = ["app", "alembic", "supabase"]
